Se modifica el valor del campo uid

+++ db/migrate/20120820043944_devise_create_users.rb
@@ -42,7 +42,7 @@ class DeviseCreateUsers < ActiveRecord::Migration
       t.string :interview
       t.string :avatar
       t.string :role
-      t.integer :uid
+      t.float :uid
       t.string :provider
       
       t.timestamps
